requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
pillow==11.2.1
numpy==2.2.4
torch
huggingface_hub>=0.23.0
git+https://github.com/microsoft/MoGe.git
